Improve gnomad formatting #7
